Influencers-dev/backend-influencers/requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
config==0.5.0.post0
Django==3.1.4
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
importlib-metadata==3.3.0
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.1
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
zipp==3.4.0
